feat: declarative database management
This patch adds a prototype of the declarative database management
feature by using a separate CRD named Database and exposing the
relative status in its section.

This new CRD is only controller by the instance manager of primary
instance.

Database deletion is triggered by a finalizer with get added
by the reconciliation loop itself. By default the retention policy is
set so "ratain", prevending the controller from deleting the database.

Closes: #4679

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>