feat: flexible data durability
This patch allows the user to relax the strict enforcement of data
durability with the new synchronous replication interface by setting
`.spec.postgresql.synchronous.dataDurability` to `preferred`.

By doing that, the operator is capping the number of required
synchronous standby replicas to the number of ready instances.

This configuration requires `standbyNamesPre` and `standbyNamesPost` to
be empty.

Closes: #5793

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>